Request to review the repa.csl
Dear volunteers, 

Thank you in advance for your time and for the review of the Research and Education Promotion Association (REPA) citation style. It worth to mention, this style will be applied for 6-journals and conferences, therefore the ISSN is ignored. Hopefully, looking to be validated successfully. 

Sincerely,
Danish M.
www.repa.jp